First commit: HTML and CSS structure!
Within this commit I have made the baseline structure for this "To Do List" project. It includes my header of the page that is broken up into a Left, Middle, and Right hand side. The next section is the main content for the to do list implementation. The next step is adding in Javascript (it will be my first attempt at adding JS to a project!).